+++ packages/common-stellar/src/project/index.ts
@@ -6,3 +6,6 @@ export * from './models';
 export * from './types';
 export * from './utils';
 export * from './versioned';
+
+import {parseStellarProjectManifest} from './load';
+export {parseStellarProjectManifest as parseProjectManifest};