+++ packages/node/src/indexer/fetch.module.ts
@@ -18,6 +18,7 @@ import {
   ProjectUpgradeService,
   InMemoryCacheService,
   SandboxService,
+  MonitorService,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { StellarApiConnection } from '../stellar/api.connection';
@@ -83,6 +84,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
         dynamicDsService: DynamicDsService,
         unfinalizedBlocks: UnfinalizedBlocksService,
         connectionPoolState: ConnectionPoolStateManager<StellarApiConnection>,
+        monitorService?: MonitorService,
       ) =>
         nodeConfig.workers
           ? new WorkerBlockDispatcherService(
@@ -98,6 +100,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
               dynamicDsService,
               unfinalizedBlocks,
               connectionPoolState,
+              monitorService,
             )
           : new BlockDispatcherService(
               apiService,
@@ -126,6 +129,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
         DynamicDsService,
         UnfinalizedBlocksService,
         ConnectionPoolStateManager,
+        MonitorService,
       ],
     },
     FetchService,
@@ -142,7 +146,8 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
       provide: 'IProjectService',
     },
     UnfinalizedBlocksService,
+    MonitorService,
   ],
-  exports: [StoreService, StoreCacheService],
+  exports: [StoreService, StoreCacheService, MonitorService, PoiService],
 })
 export class FetchModule {}