+++ packages/node/src/init.ts
@@ -3,7 +3,7 @@
 
 import { NestFactory } from '@nestjs/core';
 import { findAvailablePort, notifyUpdates } from '@subql/common';
-import { getLogger, NestLogger } from '@subql/node-core';
+import { exitWithError, getLogger, NestLogger } from '@subql/node-core';
 import { AppModule } from './app.module';
 import { FetchService } from './indexer/fetch.service';
 import { ProjectService } from './indexer/project.service';
@@ -27,12 +27,12 @@ export async function bootstrap(): Promise<void> {
 
   const port = validate(argv.port) ?? (await findAvailablePort(DEFAULT_PORT));
   if (!port) {
-    logger.error(
+    exitWithError(
       `Unable to find available port (tried ports in range (${port}..${
         port + 10
       })). Try setting a free port manually by setting the --port flag`,
+      logger,
     );
-    process.exit(1);
   }
 
   if (argv.unsafe) {
@@ -61,7 +61,6 @@ export async function bootstrap(): Promise<void> {
 
     logger.info(`Node started on port: ${port}`);
   } catch (e) {
-    logger.error(e, 'Node failed to start');
-    process.exit(1);
+    exitWithError(new Error('Node failed to start', { cause: e }), logger);
   }
 }