+++ packages/node/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Bump version with `@subql/common-stellar`, sync with `@subql/node-core`, add admin module
 
 ## [3.11.2] - 2024-05-03
 ### Fixed